init.target.rc: remove /sys/kernel/boot_adsp references in charger mode
In sun target, adsp_loader module would not be loaded hence there
is no presence of /sys/kernel/boot_adsp. This is because of ADSP
and CDSP bought out of reset by PIL automatically. Hence, waiting
for /sys/kernel/adsp_boot/boot and writing 1 to it to bring ADSP
out of reset is not needed for charger mode. Remove it.

Change-Id: Idbdf54899755f5e1fb2da772f1e6261ba1cc9a28